Initial project commit
Includes version of hello-jni that draws button on screen.
make.py is script that helps with building, especially on windows
platform.

Notes:
String and CharSequence are not supported yet, they should be first
class to be implemented.
This is just a draft, all wx classes are blank ATM

I recommend using eclipse to build this.
Made for NDK r9